Simplify folder structure to not use as many nested folders #10
