+++ client/components/Builder/WarningMessage.js
@@ -6,12 +6,9 @@ const WarningMessage = () => (
   <Container
     style={{
       width: '25%'
-    }}>
-    <Message
-      color="black"
-      header="No Pedalboard Selected"
-      content="Please select a pedalboard to continue"
-    />
+    }}
+  >
+    <Message color="black" header="No Pedalboard Selected" content="Please select a pedalboard to continue" />
     <Link style={{ marginLeft: '22px' }} to="/pedalboards">
       Go to Pedalboards Page
     </Link>
