+++ client/components/Modal/BuilderModal.js
@@ -1,19 +1,10 @@
 import React from 'react'
 import { Modal } from 'semantic-ui-react'
-import Pedals from '../../Pedals/Pedals'
+import Pedals from '../Pedals/Pedals'
 
-const BuilderModal = ({
-  closeModalHandler,
-  showModal,
-  handleClick,
-  pedalData
-}) => {
+const BuilderModal = ({ closeModalHandler, showModal, handleClick, pedalData }) => {
   const modalOutput = (
-    <Modal
-      onClick={closeModalHandler}
-      open
-      dimmer="blurring"
-      style={{ position: 'absolute', top: '13%' }}>
+    <Modal onClick={closeModalHandler} open dimmer="blurring" style={{ position: 'absolute', top: '13%' }}>
       <Modal.Header>Click a Pedal to Add to Pedalboard</Modal.Header>
       <Modal.Content image scrolling>
         <Pedals handleClick={handleClick} pedalData={pedalData} />
