--- client/components/PedalboardBuilderDisplay/PedalboardBuilderDisplay.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import { Image } from 'semantic-ui-react'
-
-const PedalboardBuilderDisplay = ({ currentPedalboard }) => {
-  return (
-    <div>
-      <Image size="massive" src={currentPedalboard.topView} centered />
-    </div>
-  )
-}
-
-export default PedalboardBuilderDisplay
