+++ client/components/Pedals/Pedal.js
